[IMP] {,helpdesk_}sale_timesheet{_enterprise,}: prevent sol from being updated on validated timesheet
before this commit, In project if company invoice by timesheet and they want to
set new sol with new prices then that sol is set for all the timesheets even if
they are validated.
prevent sol from being updated for validated timesheets and sol will only apply
on timesheets which not validated.

task-3410282

closes odoo/enterprise#43866

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>