[FIX] l10n_cl_edi: allow download of DTE attachment
The DTE attachment is currently using the model fetchmail.server with the
res_id of the journal entry when creating the DTE attachment. This will
causes issues downloading it on V16.2 resulting in an empty datas.html file.

This commit update the model to use account.move model.
It targets V14 to avoid the issues after upgrading.

opw-3360656

closes odoo/enterprise#43846

X-original-commit: d00a3771f759bd0d1366ef99b3f5a0bcc9b8cc5e
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Thomas Beckers (tbs) <tbs@odoo.com>