[ADD] l10n_nl_reports_sbr_icp: send ICP report to Dutch gov
We added the feature to send the Tax Report to the Dutch government
for our Dutch clients. This new module (which will be merged into
l10n_nl_reports in master in a new task) will allow to send the
Intra-Communautaire prestaties report (EC Sales List) the same way.

task-3187020

closes odoo/enterprise#44893

X-original-commit: 81dc618a419b88ea2d9a430fb64639d384108527
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>