[FIX] l10n_ar_reports: Tax report. Vat book (zip). Foreign partner without country.
Task Adhoc: 29354
Task latam: 1052

Steps to reproduce:

Go to runbot Odoo master enterprise and install "l10n_ar_edi" module (Argentinean Electronic Invoicing).

Take position on Argentinian company (AR) (Responsable Inscripto) .

Create a vendor bill for foreign partner. It is necessary that the partner does not have a country_id or if it does, it is needed that it does not have VAT. Also it is required the invoice to have document type (66) DESPACHO DE IMPORTACIÓN. Another requirement is that the journal use documents.

Go to Accounting / Reporting / Statement Reports / Tax Report and click on button "VAT BOOK(ZIP)", previously select the period in which the invoice created in strep 3 is included.

Current behavior:

After clicking on "VAT BOOK(ZIP)" there is a bug with traceback.

Expected behavior:

After clicking on "VAT BOOK(ZIP)" there is no bug with traceback. The user receives a message to assign a country to the partner or assign a vat to the country.

Video:
https://drive.google.com/file/d/1zBLDsvsL1xa6PY6yHwgBZdajrzkgpGfp/view

closes odoo/enterprise#43819

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>