[FIX] documents: exception while upload PDF document
When the user tries to upload a corrupted PDF in the document module, they will
face the KeyError.

Steps to produce:

- Open the document module.
- Use the following link to download the PDF document.
Link: https://github.com/py-pdf/PyPDF2/files/9066120/UTA_OSHA_3115_Fall_Protection_Training_09162021_.pdf
- Click on the Upload button to select a downloaded PDF document.
- Error will be produced

This commit addresses the traceback issue `KeyError` '/Pages' that occurs when
uploading a document in the document module. The problem arises when a PDF
document is uploaded, and the function `PdfFileReader` from the external library
`PyPDF2` is called. This function is responsible for returning the number of
pages in the PDF document and setting the `is_multipage` field to either `true`
or `false`. To resolve this problem, added a `KeyError`  Exception handler within
the existing try-catch block. By doing so, the KeyError that is causing the
traceback will be caught and handled appropriately.

Sentry-4292606596

closes odoo/enterprise#44188

X-original-commit: 347e1001b086dc61d2d69b17a5fc351556984a85
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Akshat Trivedi (aktr) <aktr@odoo.com>