[FIX] account_accountant: make sure deferral start date is before end date
In our own production database it appears that an entry had an end date
set before its start date, causing a `ZeroDivisionError` later on
in a calculation.

Normally dates like this shouldn't happen, but to prevent any stacktrace
when there is abnormal data, we check whether the end date is before
the start date of a deferral, and if so, we set both dates to the start
date to not make the reports crash (we don't know what is correct).
It is the user's responsibility to change the dates in the database.

task-3418566

closes odoo/enterprise#44239

X-original-commit: a2eb0ebdff9ea63a4ced11c9ee3685fee22376bf
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>