[FIX] account_reports: don't crash when trying to expand a groupby line without caret option
When expanding the last level of a groupby in any report, it crashed if the related model did not correspond to any caret option key either in the result of the custom handler's _caret_options_initializer() or in _caret_options_initializer_default().

This was due to the fact the controller's caretOptions() getter returned undefined, and LineName's hasCaretOptions() tried to access the 'length' field of its result.

closes odoo/enterprise#44340

X-original-commit: f3e10e9add6b4fa30327226d16cabd2e14d2a544
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>