[FIX] account_online_synchronization: perform a fetch transaction after reconnect
Reconnect is the action where the customer will try to reset the state of the connection
between his database and odoofin to 'connected'. This can happen for example
after a database failed to pay for enterprise subscription and therefore
the connection was deactivated and it needs to be reactivated using
reconnect once the customer has fix his subscription problem.
What we want is that after such a reconnect action occurs, we fetch directly the
new transactions to prevent customer to wait for the next cron iteration

closes odoo/enterprise#46030

X-original-commit: d8f5ab496a47dddf08d1e49feea920e52b5fc22e
Signed-off-by: Cedric Snauwaert <csn@odoo.com>