[FIX] l10n_ec_edi: sale person
When creating a Vendor Bill, the default Sales Team & Person of the company is
assigned to the Vendor Bill. The field isn't even displayed, it's hidden, you
can only display it with Studio and can't even change it.

In the community PR (https://github.com/odoo/odoo/pull/120628), we add a
condition on the filling of this field but by doing so we need to modify this
test.

closes odoo/enterprise#44034

Task-id: 3279233
Related: odoo/odoo#120628
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>