[FIX] account_reports: fix "Unfold All"
On the accounting reports there is an option called "Unfold All".
Selecting this option unfolds all the lines of the report.

Some lines of the report that have no children lines (neither now, nor through an expansion / unfolding)
will falsely be marked as unfolded by the "Unfold All" option.
This leads to the following:
For some of those lines a total line will be added to the report.
This total line should not appear since there are no children lines to total.

This commit fixes how / when some lines get marked as unfolded in case
the "Unfold All" option is enabled.
This also fixes the unnecessary total lines when "Unfold All" is enabled.

task-3380117

closes odoo/enterprise#44068

Signed-off-by: Quentin De Paoli <qdp@odoo.com>