[FIX] web_studio: add a test to ensure clickbot enters studio
This commit is the counter-part of odoo/odoo#127974.

I couldn't find a smarter way to ensure that the clickbot really
enters Studio. Indeed, there may be apps for which we do not enter
Studio at all, and it doesn't mean the selector is wrong. As the
click_everywhere test runs the clickbot app by app, we can't simply
assert at the end that we entered Studio at least once.

closes odoo/enterprise#43910

X-original-commit: 696b2e789c0f3f19baf75f427204d1fca4a527af
Related: odoo/odoo#128040
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>