[FIX] account_reports: improve cash flow report performance
Reason for the problem

When dealing with a large number of journal items (3 million+ AML), there is a significant slowdown in displaying the cash flow report, resulting in poor performance. The execution time for processing these moves was measured at 80s locally and 7 minutes on a hosted database, which is causing delays and impacting user experience.

Current behavior

A slowdown in loading the cash flow report with 3 million+ AML.

Expected behavior

Efficient processing of account move lines without noticeable delays.

Fix

To address this performance issue, a subquery was implemented to optimize the processing of account move lines. This subquery approach has successfully reduced the execution time from 80s to ~15s in both local/hosted db.

opw-3220665

closes odoo/enterprise#43971

Signed-off-by: William André (wan) <wan@odoo.com>