[FIX] account_reports: improve cash flow report performance
Reason for the problem

When dealing with a large number of journal items (3 million+ AML), there is a significant slowdown in displaying the cash flow report, resulting in poor performance. The execution time for processing these moves was measured at 80s locally and 7 minutes on a hosted database, which is causing delays and impacting user experience.

Current behavior

A slowdown in loading the cash flow report with 3 million+ AML.

Expected behavior

Efficient processing of account move lines without noticeable delays.

Fix

To address this performance issue, a subquery was implemented to optimize the processing of account move lines. This subquery approach has successfully reduced the execution time from 80s to ~15s in both local/hosted db.

opw-3220665

closes odoo/enterprise#44029

X-original-commit: 1db9d453ff07a40c564e718aa1c2bff66b21a1cf
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Noura El Allam (nea) <nea@odoo.com>