[FIX] project_enterprise: display project in milestone popover
Steps:
Install `sale_timesheet`.
Go to Project > All tasks > gantt.
You should see a diamond and a red line representing milestones.
Hover the diamond > a popover is displayed.
There should be 2 milestones.

Issue:
Their respective project are not displayed.

Cause:
We display it only if the context key `search_default_my_tasks` is set.
Yet in "All Tasks", it is not.

Fix:
Assuming the intention was to hide the project in the popover
when viewing the tasks of a particular project,
we now display it if the context key `default_project_id` is falsy.

task-3420545

closes odoo/enterprise#44082

X-original-commit: 401dc9194bd2436d1c4f347e6e2687d636026c3e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Audric Onockx (auon) <auon@odoo.com>