[FIX] social_facebook: update log while fetching the comment
In the _facebook_comment_fetch method, the logger is updated to use the
'warning' level instead of the 'error' level. this error might be generated
due to invalid data or having a problem with permission on Facebook.

The changes have been made to reduce the noise level in the sentry.

sentry-4242012184

closes odoo/enterprise#44094

X-original-commit: 7fc5ef6f803c0d3d6a0da9eae34d0a86c9ce9632
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>