[FIX] social_facebook: update log while fetching the comment
In the _facebook_comment_fetch method, the logger is updated to use the
'warning' level instead of the 'error' level. this error might be generated
due to invalid data or having a problem with permission on Facebook.

The changes have been made to reduce the noise level in the sentry.

sentry-4242012184

closes odoo/enterprise#42339

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>