[FIX] hr_payroll: prevent call with destroyed component
Steps to reproduce:
-------------------
- go to the form view of a Payslip;
- click on the action menu "Edit Payslip Lines";
- add a line and fill with values;

Issue:
------
A traceback appears saying the component is destroyed.

Cause:
------
When we click on "Add a line", we create a `PayslipLineField`
component for each field in the list view.

The purpose of the `fieldComponentProps` getter is to render
the component props with the record, which has a new update method.
We will redefine an update function for this record.
It will be redefined only once, via the `isPayslipLineField` flag.

We will use the first component to be created for the record,
i.e. a component for the `sequence` field.
The new update method uses `this` to make an orm call.

As a result, this will refer to the component
for the `sequence` field necessarily,
as we are defining this new update method only once.

Unfortunately, the record is in edit mode and
the `sequence` field component is not yet mounted.
 The component will therefore be destroyed
 during the "switchMode" method.

Without the `isPayslipLineField` flag,
it was "lucky" that the component existed
(the `this` was a component referring to the `total` field).

Solution:
---------
Use a component that does not refer to the `sequence` field,
so that the call can be made with a mounted component.

Note:
-----
The problem does not append for records in the `WorkedDaysField` component,
because the first field is `WorkedDaysField` and not a `sequence`.

opw-3419064

closes odoo/enterprise#43998

X-original-commit: a4c8befe98d0bde7a68568a0e9599accfe7b67b6
Signed-off-by: Kevin Baptiste <kba@odoo.com>