[FIX] sign: fix markup in green_savings_report
This commit fixes the green_savings_report template, eliminating potential
issues.

A strange call to web.html_container is removed (note that web.basic_layout
already wraps web.html_container). This call caused an entire <html> layout to
be inserted in the middle of the document.

Modal code is moved inside the call to web.basic_layout, which is the correct
place. Having the code after the call caused it to be inserted after the closing
</html> tag.

task-3380976

closes odoo/enterprise#43894

X-original-commit: 256b6a526f6e7720d62df7fc30d3cd567ffe6c6d
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Viktors Lipskis (vili) <vili@odoo.com>