[FIX] l10n_cl_edi: logger error to warning on move not found with partner
While retrieving emails from incoming customers, this issue will only be
triggered when the emails contain attachments of the .xml file type.
Subsequently, a data matching process is carried out between the email
content and the searched database. If no matching data or IDs are found,
an error is thrown.

This commit will change logger error to logger warning.

sentry-4305483242

closes odoo/enterprise#50853

X-original-commit: 0c5c5e131b86ff825c0b86431fcdb76d815bbd13
Signed-off-by: Cedric Snauwaert <csn@odoo.com>