[FIX] knowledge: properly assign default stage id
Internally, the method `_compute_stage_id` of the `knowledge.article`
model generates a hashmap mapping an article id to a stage. The hashmap
will then be used to efficiently assign a stage to an article.

When fetching a stages from that hashmap, the method uses as key a
recordset of article instead of an article id. As a result, the hashmap
does not return any entry and the article items are not be assigned to a
default stage. This commit will address that issue by ensuring that the
provided key will be an article id.

Steps to reproduce the issue:
1. Create an article with article items
2. Create a few stages and link them to the parent article
3. Read the `stage_id` value from an article item
=> The article items are not assigned to any stage

TO BE: By default, the article items should be assigned to the stage
attached to its parent having the lowest sequence number.

task-3514830

closes odoo/enterprise#47788

X-original-commit: b24aecacfbbbc2a8d02e02679d3410b0b38b8bc4
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>