[FIX] mrp_workorder{,_hr}: usable employee login on mobile
Steps to reproduce
==================

- Use a mobile sized viewport / mobile device
- Go to Manufacturing
- Click on a computer icon next to the "Plan Orders" button
- Click on the dropdown arrow
- Click on "Log In"

-> A popup should appear but nothing happens

Cause of the issue
==================

The popups are placed after the Kanban/List viewbuttons.
On mobile, this is a collapsed section.

Solution
========

For dialogs, there is a `useService("dialog")` designed to handle those
usecases. In order to use the service, the component need to be inside a
`<Dialog/>` component. For that, we can create a DialogWrapper component.
We then edit the `useConnectedEmployee` hook to use the dialog service.
With that, we need to pass the props directly. They are taken from the
previous template.

opw-3418639

closes odoo/enterprise#44119

X-original-commit: ac3ecd4b42e067cd8b3e15949a66696e7ed421e2
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>