[IMP] knowledge: render templates insertion collaborative-friendly
Ensure that the `WysiwygHelper` is correctly displayed even after applying a
bunch of steps coming from the collaboration (`historyExternalSteps`).

Ensure that the backend does not write on the body of an existing article, since
this will either be non-effective if a collaborator is editing the article
currently and succeeds to save after the write, or it risks breaking the
collaboration in case the user which initiated the write was in the
collaboration, and succeeds in reloading the view with the new body value after
the write, since in that case the editor does not have the time to write
collaborative steps id in the first node resulting in a `stale` situation where
the editor thinks the new body is a stale version of the document that is still
being edited with its old value by the other collaborators.

The solution is to allow `apply_template_on_article` to do all non-collaborative
write, and return the body value without writing it. The client which applied
the template will then create a collaborative step with that new value to share
it with collaborators, using `wysiwyg.setValue`.

task-3378266

X-original-commit: odoo/odoo@f705e9c20715d370da071ecb4ef8f0e6b8703d79
Part-of: odoo/enterprise#44317