[IMP] knowledge: allow collaborative mode for portal users
ICE servers are a mail feature, and the `editor` uses a hack to check if it is
installed by checking `getSession().notification_type`. The issue is that this
value is not present for portal users that should be able to use the
collaborative feature for Knowledge.

Since `mail` is a dependency of Knowledge, we can override the hack to let the
editor know that the ICE servers are available as long as Knowledge is
installed.

task-3378266

X-original-commit: odoo/odoo@d6e4cf1af4523e3efcc6c9b63ea2549fabae00b6
Part-of: odoo/enterprise#44317