[FIX] appointment: fix linked resource availability
Don't take into account linked resources that don't have remaining
capacity or that are not part of the initial resources available.

This was leading to some errors when validating a time slot selected
because some resources with no remaining capacity where added from
resources they were linked to.

task-3424035

X-original-commit: 36f4db270bf58337fe3e8a610c2f04109fb90548
Part-of: odoo/enterprise#49845