[FIX] mrp_workorder: setup test to work with employee without request
The employee loged on mrp is done via the request.session. However in
test we don't have it and we are using the super user. Modify the
test to take the admin user and select his employee by default when
there is no request

closes odoo/enterprise#43610

Related: odoo/odoo#127250
Related: odoo/upgrade#4938
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>