[IMP] web_map: navigate to the address instead of coordinate
This commit allows to avoid the lost of precision when clicking to
navigate on the map view.

Before this commit, to display the popup on the map view, the address
is converted to longitude/latitude. And when you click on the 'navigate'
we use this longitude/latitude to specify the destination to google, and
google will revert these coordinates to an address that could not be the
same as the original.

After this commit, we send the address to Google as destination when the
user clicks on the 'Navigate' button.

opw-3346406

closes odoo/enterprise#44060

Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>