[FIX] *: Unwanted traceback on "Bank" journal report's first opening
*: account_reports

Fixes a bug that occurred when opening the "Bank" journal report for the first
time after a refresh or changing the filters in specific cases.

Steps to reproduce :
1 - Go to Journal Reports
2 - Click on "Bank (BNK1)" to unfold the journal
3 - A traceback should pop

This is fixed by modifying the behavior of the unfold function. When the total line
is added to the lines that will be displayed, a check is done to verify that it exists.
If said line does not exist, then it is not added so the traceback doesn't pop.

task-3424016

closes odoo/enterprise#45356

X-original-commit: f0811c18b200e1ab82a6fd5424f30d9c2e7db837
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>