[IMP] knowledge: add emojipicker to item kanban quick create
Purpose:
-------
Allow users to set and modify an icon from the quick create cards in
knowledge's item kanbans.
The quick create cards now show an icon placeholder when being hovered,
that changes into a random icon when clicked.
This icon can be clicked again to select an icon using the emojipicker.

Task-3291762

Part-of: odoo/enterprise#44072