[REM] website_sale_dashboard: remove website sale community dashboard
This dashboard depends on the website dashboard which is to be re-done
in OWL. Instead of keeping this dashboard as a separate component, it
will be re-done in the Dashboard App (see task-3222991).

The enterprise module referenced the community dashboard to hide it,
this is no longer needed.

task-3164163

closes odoo/enterprise#44087

Related: odoo/odoo#112819
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>