[IMP] web_studio: handle edited reportId in URL
Before this commit, when being in the report editor, the url did not
mention the report id that was being edited. This meant that a reload of the
browser made the user land back onto the kanban view to select a report.

After this commit, we are able to load the report editor from the url

Part-of: odoo/enterprise#43978