[FIX] industry_fsm_sale: hide SOL field on non-billable projects
When the SOL field was moved from the 'Invoicing' tab to the 'Settings' tab
in the project form view, the visibility of the field was not properly updated.

An override in `industry_fsm_sale` erroneously removes the condition that the
project has to be billable and have a partner set.

This commit adds this condition to the override.

closes odoo/enterprise#44235

X-original-commit: b2e61a71a4ec5a253ef0596977cf5bc9a5673fb0
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>