[FIX] helpdesk: ignore default_team_id in partner view
Steps to reproduce:

- Install CRM, Helpdesk.
- Go to Settings > Discuss > Check `Custom Email Servers`,
Alias Domain = `@example.com`.
- CRM > Config > Sales teams > New > Enter Name = Team1 > Save.
- Helpdesk > Create a new customer and record some tickets for them,
or use an existing customer (e.g., Azure Interior).
- CRM > Pipeline > New > Enter Contact = Azure Interior > Edit.
	- In the Extra Information tab, set Sales Team = Team1 > Save.
	- From the opportunity view, open the Azure Interior record
	through the internal link.
- From the customer's profile, click on the Tickets stat button.

Issue:

Traceback: `Record does not exist or has been deleted.` This error
occurs because the sales team of the opportunity is passed in the
context as `default_team_id`. When clicking the stat button, we search
for a helpdesk team using the same context variable, `default_team_id`.
Since there is no helpdesk team with that id, the traceback appears.

Solution:

When clicking on the tickets stat button, our aim is to display all
helpdesk tickets related to that customer, and the helpdesk team is not
relevant in this context. Therefore, we can set `default_team_id` to
False before displaying the tickets, and this change will not affect
the results.

opw-3371767

closes odoo/enterprise#44896

X-original-commit: 619c00eb5fdfd24ea77cb94d26557d6eea83c384
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>