[IMP] appointment: Add leaves management button
Before this adding resource leaves wasn't practical
when just working with appointments.

We add a button and wizard on the gantt view of
calendar events when grouping by resource. This
enables users to simply add leaves on resources
from the main appointment resource management view.

We also add a settings menu directly to resource leaves
for users to remove and edit leaves if necessary.

Additionally, resources did not respect their calendar
availabilities prior to this commit. We add some checks
for slot generation to make sure resources are marked
unavailable according to their calendar.

task-3101589

Part-of: odoo/enterprise#37312