[FIX] account_asset: error message for asset disposal before lock date
When the user tries to dispose an asset before the defined user lock
date
the following User Error appears:
"The remaining value on the last depreciation line must be 0".
It may be unclear to the user that the lock date is responsible for this problem.

After this commit the error message makes it clear
that the lock date is the reason why the disposal is not possible.

Steps to reproduce the User Error:
  1. Set an "All Users Lock Date"
     (Accounting (App) > Accounting (Menu) > Lock Dates)
     E.g. 28 June 2023
  2. Create an Asset with a non-0 value.
     (Accounting (App) > Accounting (Menu) > Assets)
     E.g. choose
     Original value: 100
     Acquisition date: 1 June 2023
  3. Click "Modify Depreciation" on the (confirmed) asset.
     The "Dispose" action should be selected.
     Choose a date between the acquisition date and the lock date
     e.g. 14 June 2023
     and click "Dispose"
  4. A User Error appears:
     "The remaining value on the last depreciation line must be 0".

task-3377235

closes odoo/enterprise#44254

Signed-off-by: Quentin De Paoli <qdp@odoo.com>