[FIX] hr_payroll: use correct report(s) on Print
The automatic PDF generation would use the report(s) defined in
`_get_pdf_reports()`, however the "Print" button would only use the
report defined on the Payroll Structure.

This meant that using the "Print" button on an Holiday Attest (Belgium
localization) payslip would only print the payslip itself and not the
attest.

closes odoo/enterprise#44591

X-original-commit: 248852c4bb538ec84edbf8545c7a04b413779817
Signed-off-by: Kevin Baptiste <kba@odoo.com>