[ADD] l10n_ar_edi: Added new logic for new units of measure
Before this PR the users were not able to generate electronic invoices
to the FEX and FBE webservices that have negative lines (For example:
downpayment, bonuses, coupons, global discount lines).

This because the WS only accept negative lines if we sent some specific
information within the xml. This PR make it work by evaluationg the
negative lines, deducing the reason behind the line and sending to the
webserve the data adapted so they can report the line correctly

resolves opw-3572679
latam task 1064
adhoc task: 32083

closes odoo/enterprise#44218

Signed-off-by: Josse Colpaert <jco@odoo.com>