[FIX] web_grid: hide custom button if invisible = true
Before this commit, when the custom button should be invisible when
a context is not there or falsy, the custom button is always displayed
because the `invisible` modifier is not used to conditionally display
the custom button.

This commit adds a visibility condition to display the custom if
invisible modifier is falsy, otherwise the custom button is not
rendered.

task-3222154

Part-of: odoo/enterprise#44599