[FIX] project_enterprise: hide current user if filtered by user_ids
Currently, in the project.task Gantt view, the row of the current user
is always displayed, even if empty and if a search is performed on the
assignees of tasks. This behavior is fixed to hide this row should if a
filter is applied on user_ids.

Steps
=====
- Install project_enterprise
- Go in the Gantt view of a project
- Search for an assignee in the search bar
- The current user row is always visible

Issue
=====
The current user row is always visible, even if the filter on user_ids
does not match the current user.

Cause
=====
The group expand method on user_ids is designed to always add the
current user.

Fix
===
A check is performed on the domain to see if a filter is applied on
user_ids. If it's the case, the current user is not automatically added
to the resulting group.

task-3429600

closes odoo/enterprise#45326

X-original-commit: 51e5d6fdc507d91254431bfc9583c6fc0ae66faf
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>