[FIX] social_facebook: filter out false facebook_account_id
When the user opens Social Marketing app and `facebook_account_id` is set to
`False` the following error will be generated.

Error: A traceback appears: `TypeError: Cannot mix str and bytes arguments (got
('https://graph.facebook.com/v17.0/', False))`

In the neutralized database, a SQL query is fired which is setting
`facebook_account_id` to NULL. The compute method `_compute_statistics` is using
`url_join` which is joining the Graph API of  Facebook with
`facebook_account_id` but as `facebook_account_id` is NULL so `url_join` is
throwing a `TypeError`.

To resolve this issue, the old code loops through all the `facebook_accounts`
without any specific criteria. The new code filters the `facebook_accounts`
collection and only iterates over the accounts that have a non-empty
`facebook_account_id`.

Sentry-4312545705

closes odoo/enterprise#44253

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>