[IMP] account_accountant: Bank reco: improved exchange diff audit
When reconciling through the bank reco widget where an exchange difference
was necessary, we used to add an aml with that exchange difference to the
statement's related move.
In consequence, the exchange difference wasn't linked to a different move
like we do elsewhere in Odoo, the exchange difference aml was also not
linked to the matching number which made the audit/tracking harder.

task-3415644

closes odoo/enterprise#44274

Signed-off-by: Laurent Smet (las) <las@odoo.com>