[FIX] web_studio: fix editor link
Block blur event when a modal is open.

When the user open a link modal in the odoo editor
(via the power box command) a blur event is triggered
and force the save and reload of the  editor.

We block the blur event if a modal is open,
to be sure the user can add a link without issue.

When the link modal closes, the focus is put back
in the editor, ensuring the blur event will be triggered
next time the user leave the editable document.

task-3058355

closes odoo/enterprise#47869

X-original-commit: 01502fd9fc304feaf6b8d3373f5ffb3cb6bc4b2d
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>