[FIX] documents: thumbnails changes accordingly when changing pdf file
Steps to reproduce:

- Go to document app.
- Select any existing PDF document.
- Use the replace action to change the PDF.
- The thumbnail is broken (until changing it again).

Solution:

Since we really only get the PDFs thumbnails from the js side,
specifically from the `documents_pdf_thumbnail_service.js` we won't be
able to generate the thumbnail from the `_compute_thumbnail`. This means
that the issue is the conflict between the backend generation of the
thumbnail and the js, so for the PDFs we will only need the js part to
generate and set the proper status for the thumbnails.

opw-3371883

closes odoo/enterprise#46113

X-original-commit: b6509200af1ba207cef1ccb06d7b0f23778dacb9
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>