[FIX] planning: speed up `_init_column`
Description:
During the installation of the module `planning`, we have an
override of `_init_column` for the `employee_token`. We were using
the psycopg's `executemany` which is just a syntactical sugar for a
loop over an `execute`. So we trigger 1 query (which incurs a
network cost when deployed and the database is not on the same
machine) per row, currently we have 20 queries.

Solution:
We can use SQL syntax:
```sql
UPDATE ... SET ... FROM (VALUES ...) WHERE ...
```
which does 1 update for all rows in one go.

Implementation details:
We could implement this query directly with a simple execute, but it
requires a lot of manual construction with the help of `sql.Composed`,
because of necessity to insert manually `sql.Literal`
used for the tuples of the `VALUES`.
Psycopg2 exposes a function called `execute_values` which is just
for this case and makes to code way more legible.

task-3430406

closes odoo/enterprise#44268

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>