[FIX] event_barcode: speed up `_init_column`
Description:
During the installation of the module `event_barcode`, we have an
override of `_init_column` for the `barcode`. We were using
the psycopg's `executemany` which is just a syntactical sugar for a
loop over an `execute`. So we trigger 1 query (which incurs a
network cost when deployed and the database is not on the same
machine) per row, currently we have 27 queries.

Solution:
We can use SQL syntax:
```sql
UPDATE ... SET ... FROM (VALUES ...) WHERE ...
```
which does 1 update for all rows in one go.

Implementation details:
We could implement this query directly with a simple execute, but it
requires a lot of manual construction with the help of `sql.
Composed`, because of necessity to insert manually `sql.Literal`
used for the tuples of the `VALUES`.
Psycopg2 exposes a function called `execute_values` which is just
for this case and makes to code way more legible.
We also generate barcodes only for registrations linked to event that
are ongoing or in the future, to improve the speed of the UPDATE itself
(we update less rows).

Reference:
task-3430406

closes odoo/enterprise#44271

Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>